% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cas_conv}
\alias{cas_conv}
\title{Utility functions for the standartox package
Convert CAS betwenn hyphen and hyphenless form}
\usage{
cas_conv(cas)
}
\description{
Utility functions for the standartox package
Convert CAS betwenn hyphen and hyphenless form
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
\keyword{internal}
