% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Function to flag outliers}
\usage{
flag_outliers(x, lim = 1.5, na.rm = TRUE, ...)
}
\description{
Function to flag outliers
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
\keyword{internal}
