% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculate geometric mean}
\usage{
gm_mean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\description{
Calculate geometric mean
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
\keyword{internal}
