% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{standartox}
\alias{standartox}
\title{Retrieve Standartox toxicity values}
\usage{
standartox(version_epa = NULL, cas = NULL, concentration_type = NULL,
  chemical_class = NULL, taxa = NULL, habitat = NULL,
  region = NULL, duration = c(24, 48), effect = NULL,
  endpoint = NULL, aggregate = NULL, ...)
}
\arguments{
\item{cas}{character, integer; Limit data base query to specific CAS numbers, multiple entries possible (e.g. 1071-83-6, 1071836), NULL (default)}

\item{concentration_type}{character; Limit data base query to specific concentration types, can be one of NULL (default), 'active ingredient', 'formulation', 'total', 'not reported', 'unionized', 'dissolved', 'labile'. See \url{https://cfpub.epa.gov/ecotox/pdf/codeappendix.pdf} p.4}

\item{chemical_class}{character; Limit data base query to specific chemical classes, multiple entries possible, NULL (default)}

\item{taxa}{character; Limit data base query to specific taxa, multiple entries possible, NULL (default)}

\item{habitat}{character; Limit data base query to specific organism habitats, can be one of NULL (default) 'marine', 'brackish', 'freshwater'}

\item{region}{character; Limit data base query to organisms occurring in specific regions, can be one of NULL (default) 'africa', 'america_north', 'america_south', 'asia', 'europe', 'oceania'}

\item{duration}{integer vector of length two; Limit data base query to specific test durations (hours) (e.g. c(24, 48))}

\item{effect}{character; Limit data base query to specific effect groups, multiple entries possible (e.g. 'MOR', 'ITX', 'GRO'). See \url{https://cfpub.epa.gov/ecotox/pdf/codeappendix.pdf} p.95}

\item{endpoint}{character; Choose endypoint type, must be one of 'XX50' (default), 'NOEX', 'LOEX'}

\item{aggregate}{character; Choose aggregation method, can be one of 'min', 'med', 'gmn' (default), 'max'}

\item{...}{currently not used}

\item{version}{integer; Choose the version of the EPA Ecotox on which Standartox is based on. NULL (default) accesses the most recent version}
}
\value{
Returns a list of three data.tables (filtered data base query results, aggregated data base query results, meta information)
}
\description{
Retrieve toxicity values from the Standartox data base \url{http://standartox.uni-landau.de/}.
}
\examples{
\donttest{
# might fail if API is not available
standartox('1071-83-6')
standartox('1071-83-6', duration = c(48, 120))
standartox('1071-83-6', duration = c(48, 120), endpoint = 'XX50', aggregate = 'md')
}

}
\author{
Andreas Scharmueller \email{andschar@proton.com}
}
