% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_aggregate}
\alias{stx_aggregate}
\title{Function to aggregate filtered test results}
\usage{
stx_aggregate(dat = NULL)
}
\description{
Function to aggregate filtered test results
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
\keyword{internal}
