% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_chem}
\alias{stx_chem}
\title{Retrieve chemical data}
\usage{
stx_chem(silent = FALSE, stx_dir = file.path(tempdir(), "standartox"))
}
\arguments{
\item{silent}{logical; If TRUE, suppresses messages. Default is FALSE.}

\item{stx_dir}{character; Directory to which the chemical information should be downloaded. Default is a temporary directory.}
}
\value{
Returns a data.table containing informaiton on chemicals in Standartox.
}
\description{
Retrieve data on all chemicals in Standartox.
}
\examples{
\donttest{
# might fail if there is no internet connection or Zenodo.org not not available
# basic function call
df = stx_chem()

# to get verbose output from the function
df = stx_chem(silent = FALSE)

# to specify a directory to which the chemical information should be downloaded
df = stx_chem(silent = FALSE, stx_dir = "~/tmp")
# This will create a directory under ~/tmp and download the respective standartox file to that directory.
# The files are then permanently stored in that directory and can be directly read when restarting your R session.
}

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}

Hannes Reinwald
}
