% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_data}
\alias{stx_data}
\title{Retrieve Standartox toxicity values}
\usage{
stx_data(silent = FALSE, stx_dir = file.path(tempdir(), "standartox"))
}
\arguments{
\item{silent}{logical; If TRUE, suppresses messages. Default is FALSE.}

\item{stx_dir}{character; Directory to which the catalog should be downloaded. Default is a temporary directory.}
}
\value{
Returns a data.table.
}
\description{
Retrieve a data.table containing the Standartox toxicity data
}
\examples{
\donttest{
# might fail if there is no internet connection or Zenodo.org not not available
# basic function call

dt = stx_data()

}
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}

Hannes Reinwald
}
