% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stx_message}
\alias{stx_message}
\title{Compose query message}
\usage{
stx_message(body)
}
\description{
Compose query message
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
\keyword{internal}
