% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_catalog}
\alias{stx_catalog}
\title{Retrieve data catalog}
\usage{
stx_catalog(silent = FALSE, dir_out = file.path(tempdir(), "standartox"))
}
\arguments{
\item{silent}{logical; If TRUE, suppresses messages. Default is FALSE.}

\item{dir_out}{character; Directory to which the catalog should be downloaded. Default is a temporary directory.}
}
\value{
Returns a list of data.frames containing information on data base variables
}
\description{
Retrieve a data catalog for all variables (and their values) that can be retrieved with stx_query()
}
\examples{
\donttest{
# might fail if there is no internet connection or Zenodo.org not not available
# basic function call
l = stx_catalog()

# to get verbose output from the function
l = stx_catalog(silent = FALSE)

# to specify a directory to which the catalog should be downloaded
l = stx_catalog(silent = FALSE, dir_out = "~/tmp")
# This will create a directory under ~/tmp and download the catalog.rds file to that directory.
# The files are then permanently stored in that directory and can be directly read when restarting your R session.
}
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}

Hannes Reinwald
}
