% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_catalog}
\alias{stx_catalog}
\title{Retrieve data catalog}
\usage{
stx_catalog(vers = NULL)
}
\arguments{
\item{vers}{integer; Choose the version of the EPA Ecotox on which Standartox is based on. NULL (default) accesses the most recent version}
}
\value{
Returns a list of data.frames containing information on data base variables
}
\description{
Retrieve a data catalog for all variables (and their values) that can be retrieved with stx_query()
}
\examples{
\donttest{
# might fail if API is not available
stx_catalog()
}

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
