% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_chem}
\alias{stx_chem}
\title{Retrieve chemical data}
\usage{
stx_chem()
}
\value{
Returns a data.table containing informaiton on chemicals in Standartox.
}
\description{
Retrieve data on all chemicals in Standartox.
}
\examples{
\donttest{
# might fail if API is not available
stx_chem()
}

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
