% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_meta}
\alias{stx_meta}
\title{Retrieve meta data}
\usage{
stx_meta(silent = FALSE, dir_out = file.path(tempdir(), "standartox"))
}
\arguments{
\item{silent}{logical; If TRUE, suppresses messages. Default is FALSE.}

\item{dir_out}{character; Directory to which the meta information should be downloaded. Default is a temporary directory.#'}
}
\value{
Returns a data.table containing meta informaiton on Standartox.
}
\description{
Retrieve meta data
}
\examples{
\donttest{
# might fail if there is no internet connection or Zenodo.org not not available
# basic function call
df = stx_meta()

# to get verbose output from the function
df = stx_meta(silent = FALSE)

# to specify a directory to which the taxa information should be downloaded
df = stx_meta(silent = FALSE, dir_out = "~/tmp")
# This will create a directory under ~/tmp and download the respective standartox file to that directory.
# The files are then permanently stored in that directory and can be directly read when restarting your R session.
}

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}

Hannes Reinwald
}
