% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_query}
\alias{stx_query}
\title{Query Standartox Toxicity Data}
\usage{
stx_query(
  cas_number = NULL,
  endpoint_group = c("XX50", "NOEX", "LOEX"),
  exposure = NULL,
  effect = NULL,
  duration = c(0, Inf),
  duration_unit = "h",
  concentration_unit = NULL,
  concentration_type = NULL,
  tax_columns = c("group", "taxon", "genus", "family"),
  tax_genus = NULL,
  tax_family = NULL,
  tax_order = NULL,
  tax_class = NULL,
  tax_group = NULL,
  include_reference = FALSE,
  rm_NR = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cas_number}{character; Optional. Vector of CAS numbers to filter chemicals (e.g. \code{c("1071-83-6","63-25-2","138261-41-3")}). Default is \code{NULL} (no filtering).}

\item{endpoint_group}{character; Optional. Endpoint group(s) to filter results. All possible endpoint groups can be checked via \code{stx_catalog()$endpoint_group}. Default is \code{c("XX50", "NOEX", "LOEX")}.}

\item{exposure}{character; Optional. Vector of exposure types (e.g. \code{"aquatic"}). All possible exposure types can be checked via \code{stx_catalog()$exposure}. Default is \code{NULL}.}

\item{effect}{character; Optional. Vector of effect types (e.g. \code{"Mortality", "Growth"}). All possible effect types can be checked via \code{stx_catalog()$effect}. Default is \code{NULL}.}

\item{duration}{numeric; Optional. Numeric vector of length two specifying minimum and maximum test duration (in hours), e.g. \code{c(0, 48)}. Default is \code{c(0, Inf)}.}

\item{duration_unit}{character; Optional. Filter by duration unit (e.g. \code{"h"} for hours). All possible duration units can be checked via \code{stx_catalog()$duration_unit}. Set to \code{NULL} to keep all. Default is \code{"h"}.}

\item{concentration_unit}{character; Optional. Filter by concentration unit (e.g. \code{"g/l"}). All possible concentration units can be checked via \code{stx_catalog()$concentration_unit}. Default is \code{NULL}.}

\item{concentration_type}{character; Optional. Filter by concentration type (e.g. \code{"active ingredient"}). All possible concentration types can be checked via \code{stx_catalog()$concentration_type}. Default is \code{NULL}.}

\item{tax_columns}{character; Columns of taxonomic information to append to results. All possible columns can be checked via \code{colnames(stx_taxa())}. Default is \code{c("group", "taxon", "genus", "family")}.}

\item{tax_genus}{character; Optional. Filter by genus. All possible genera can be checked via \code{stx_catalog()$genus}. Default is \code{NULL}.}

\item{tax_family}{character; Optional. Filter by family. All possible families can be checked via \code{stx_catalog()$family}. Default is \code{NULL}.}

\item{tax_order}{character; Optional. Filter by order. All possible orders can be checked via \code{stx_catalog()$order}. Default is \code{NULL}.}

\item{tax_class}{character; Optional. Filter by class. All possible classes can be checked via \code{stx_catalog()$class}. Default is \code{NULL}.}

\item{tax_group}{character; Optional. Filter by one or more ecotoxicological groups. Possible values are \code{"invertebrate"}, \code{"plant"}, \code{"fish"}, \code{"fungi"}, \code{"algae"}, \code{"aves"}, \code{"amphibia"}, \code{"mammalia"}, \code{"reptilia"}, \code{"macrophyte"}. All possible ecotox groups can be checked via \code{stx_catalog()$group}. Multiple entries possible. Default is \code{NULL}.}

\item{include_reference}{logical; If \code{TRUE}, append reference information. Default is \code{FALSE}.}

\item{rm_NR}{logical; If \code{TRUE}, remove rows with "NR" (not reported) values. Default is \code{TRUE}.}

\item{verbose}{logical; If \code{TRUE}, print progress messages. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{stx_download}.}
}
\value{
Returns a \code{data.table} with filtered Standartox toxicity data.
}
\description{
Retrieve and filter toxicity data from the Standartox database (\url{https://doi.org/10.5281/zenodo.3785030}) using chemical, experimental, and taxonomic criteria.
}
\examples{
\donttest{

Basic stx_query() call: Will return results filtered for default endpoint_group = c("XX50", "NOEX", "LOEX") and duration_unit = "h"
stx_query(verbose = T)

# If you wish to filter for different endpoint groups, you can specify them in the query.
stx_catalog()$endpoint_group # to view available endpoint groups
stx_query(endpoint_group = c("Bioconc","MATC","MCIG"))

# Query for a specific CAS number, endpoint group, and tax group(s)
stx_query(
  cas_number = "1071-83-6",
  endpoint_group = c("NOEX","LOEX"),
  duration = c(0, 120),
  ecotox_group = c("invertebrate", "fish", "algae"),
)

# get ALL LC50 values for 96 - 120 h of exposure for zebra fish (Danio rerio)
stx_query(
  endpoint_group = "XX50",
  duration = c(96, 120),
  effect = "mortality",
  concentration_unit = "g/l",
  concentration_type = "active ingredient",
  tax_genus = "Danio",
  include_reference = TRUE
)

# get ALL LC50 values for 24 - 48 h of exposure for the family of Daphniidae
stx_query(
  endpoint_group = "XX50",
  duration = c(24, 48),
  effect = "mortality",
  concentration_unit = "g/l",
  concentration_type = "active ingredient",
  tax_family = "Daphniidae"
  include_reference = TRUE
)
}

}
\author{
Hannes Reinwald
}
