% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standartox.R
\name{stx_taxa}
\alias{stx_taxa}
\title{Retrieve taxa data}
\usage{
stx_taxa()
}
\value{
Returns a data.table containing informaiton on taxa in Standartox.
}
\description{
Retrieve data on all taxa in Standartox.
}
\examples{
\donttest{
# might fail if API is not available
stx_taxa()
}

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
